requirements.txt
uvicorn~=0.17.6
fastapi~=0.75.0
pydantic~=1.9.0
Jinja2~=3.1.1
PyYAML~=6.0
oauth2client~=4.1.3
google-api-python-client~=2.43.0
boto3~=1.21.37